question_id 参数 为 第几道提

+++ Application/Common/Common/function.php
@@ -119,7 +119,7 @@
         $arrValues = array_values($arr);
         $count = ($fx + $dx + $pd) * 0.7;
 
-		$count = is_int($count) ? $count : floor($count);
+		$count = is_int($count) ? $count : (int)floor($count);
 
         $arrFirstValue = 0;
         $arrSecondValue = 0;
